Correção dos exercicios e inicio de ASP.NET MVC

+++ 01.Fiap.HelloWorld.Console.UI/Exercicio-01/Model/IContaInvestimento.cs
@@ -8,6 +8,6 @@ namespace Exercicio_01.Model
 {
     interface IContaInvestimento
     {
-        public double CalculaRetornoInvestimento();
+        public decimal CalculaRetornoInvestimento();
     }
 }
